Add types/homepage.d.ts matching the API response
